feat: DifyEditorWindow SSEストリーミング表示機能を修正
SSEストリーミングが表示されない問題を解決：
- DifyHttpAdapter: JSON property名をDify API仕様に準拠
  - eventType → event
  - conversationId → conversation_id
  - messageId → message_id
- Mock/Real両モードでSSEストリーミング動作確認
- SSE録画ファイルパス設定UI追加
- UnityWebRequestHttpClient実装追加（Real通信用）
- MockHttpClient実装追加（SSERecordings再生用）
- DifyInstaller追加（MonoBehaviour標準パターン）

これにより、DifyEditorWindowでリアルタイムに
ストリーミングレスポンスが表示されるように修正

+++ aituber_3d/Assets/Scripts/Services/Dify/Presentation/Controllers/DifyController.cs
@@ -3,6 +3,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using AiTuber.Services.Dify.Application.UseCases;
 using AiTuber.Services.Dify.Domain.Entities;
+using AiTuber.Services.Dify.Infrastructure.Http;
 using UnityEngine;
 
 #nullable enable
@@ -67,7 +68,7 @@ namespace AiTuber.Services.Dify.Presentation.Controllers
                     response.TextResponse,
                     response.ConversationId,
                     response.MessageId,
-                    response.ProcessingTimeMs);
+                    (int)response.ProcessingTimeMs);
             }
             catch (OperationCanceledException)
             {
@@ -121,7 +122,7 @@ namespace AiTuber.Services.Dify.Presentation.Controllers
                     response.TextResponse,
                     response.ConversationId,
                     response.MessageId,
-                    response.ProcessingTimeMs);
+                    (int)response.ProcessingTimeMs);
             }
             catch (OperationCanceledException)
             {
